Refactoring: Split stopwatch class our into header and source instead of just a hpp file
